+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^13.5.0",
+    "classnames": "^2.5.1",
     "html-react-parser": "^5.2.6",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
