+++ src/Home/Home.module.css
@@ -1,4 +1,4 @@
-.blog__list{
+.blogList {
   display: flex;
   flex-direction: column;
   gap: 32px;
@@ -7,35 +7,35 @@
   padding: 0 1rem;
 }
 
-.blog__item{
+.blogItem {
   list-style: none;
   border: 1px solid #ccc;
   padding: 1rem;
-      
+
 }
 
-.blog__link{
+.blogLink {
   color: #333;
   text-decoration: none;
 }
 
-.blog__meta{
+.blogMeta {
   display: flex;
   justify-content: space-between;
   align-items: center;
 }
 
-.blog__date{
+.blogDate {
   color: #888;
   font-size: .8rem;
 }
 
-.blog__categories{  
+.blogCategories {
   display: flex;
   gap: 10px;
 }
 
-.blog__category{
+.blogCategory {
   border: 1px solid #06c;
   border-radius: .2rem;
   color: #06c;
@@ -44,13 +44,13 @@
   list-style: none;
 }
 
-.blog__title{
+.blogTitle {
   font-size: 1.5rem;
   margin-bottom: 1rem;
   margin-top: .5rem;
 }
 
-.blog__excerpt{
+.blogExcerpt {
   -webkit-line-clamp: 2;
   -webkit-box-orient: vertical;
   display: -webkit-box;
