+++ src/components/Header/Header.module.css
@@ -5,7 +5,7 @@ header {
   padding: 24px;
 }
 
-.header__list {
+.headerList {
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -13,11 +13,11 @@ header {
   padding: 0;
 }
 
-.header__list-item {
+.headerItem {
   list-style: none;
 }
 
-.header__link {
+.headerLink {
   color: #fff;
   text-decoration: none;
 }
